move tables & views into /tabulars
