+++ crates/iceberg-catalog/src/catalog/tabulars/mod.rs
@@ -1,0 +1,2 @@
+mod tables;
+pub(crate) mod views;
