+++ crates/iceberg-catalog/src/catalog/tabulars/views/list.rs
